chore: move render field to component (#17084)
## Description

As part of the action selector refactor, we are making the following changes - 
- moving renderField to its own component called Field
- created FieldConfig to hold all the configurations of these functions fields: This config will slowly be built up to hold all the miscellaneous configurations which is currently scattered throughout different files
- added types for the props
- moved FieldConfig to its own type

Fixes #16934 #16936

## Type of change

- Refactor

## How Has This Been Tested?

- manually
- Jest test cases

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
