+++ app/client/src/components/editorComponents/ActionCreator/constants.ts
@@ -61,6 +61,7 @@ export const ViewTypes = {
   TEXT_VIEW: "TEXT_VIEW",
   BOOL_VIEW: "BOOL_VIEW",
   TAB_VIEW: "TAB_VIEW",
+  NO_VIEW: "NO_VIEW",
 };
 
 export const NAVIGATE_TO_TAB_OPTIONS = {
@@ -70,7 +71,6 @@ export const NAVIGATE_TO_TAB_OPTIONS = {
 
 export enum FieldType {
   ACTION_SELECTOR_FIELD = "ACTION_SELECTOR_FIELD",
-  JS_ACTION_SELECTOR_FIELD = "JS_ACTION_SELECTOR_FIELD",
   ON_SUCCESS_FIELD = "ON_SUCCESS_FIELD",
   ON_ERROR_FIELD = "ON_ERROR_FIELD",
   SHOW_MODAL_FIELD = "SHOW_MODAL_FIELD",
